Move presenters in their own subpackages for manager & publisher
