+++ publisher/presenter/html/testdata/objects.html.sample
